Enemies Seperated from the Combat Manager
