+++ Assets/Prefabs/Combat/Combat Manager.prefab
@@ -20,6 +20,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4024941165589956}
   - component: {fileID: 114479076340579032}
+  - component: {fileID: 114747481958685648}
   m_Layer: 0
   m_Name: Combat Manager
   m_TagString: Untagged
@@ -94,9 +95,16 @@ MonoBehaviour:
     m_TypeName: UnityEngine.Events.UnityEvent, UnityEngine, Version=0.0.0.0, Culture=neutral,
       PublicKeyToken=null
   m_GridMono: {fileID: 0}
-  m_CurrentEnemy: {fileID: 0}
-  doCombat: 1
+--- !u!114 &114747481958685648
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1561642928282028}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 993fc86c1861de943853adc63c2b23c7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   enemyPadding: 1
-  enemyPrefabList:
-  - {fileID: 1978000353783988, guid: 05fc6f9ca8e594e429fdc6e7984fd2c6, type: 2}
-  - {fileID: 1978000353783988, guid: e4e6db6db2938154692b0ece784cb965, type: 2}
+  doCombat: 1
