+++ Assets/Scripts/Combat/EnemyManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 993fc86c1861de943853adc63c2b23c7
+timeCreated: 1490807302
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
